requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohttp=3.8.1=pypi_0
aiosignal=1.2.0=pypi_0
alembic=1.8.1=pypi_0
asgiref=3.5.2=pypi_0
async-timeout=4.0.2=pypi_0
attrs=21.4.0=pypi_0
backports-zoneinfo=0.2.1=pypi_0
ca-certificates=2022.07.19=haa95532_0
certifi=2022.6.15=py38haa95532_0
cffi=1.15.1=pypi_0
charset-normalizer=2.1.0=pypi_0
click=8.1.3=pypi_0
colorama=0.4.5=pypi_0
cryptography=38.0.1=pypi_0
defusedxml=0.7.1=pypi_0
django=4.0.6=pypi_0
django-allauth=0.51.0=pypi_0
flask=2.1.3=pypi_0
flask-migrate=3.1.0=pypi_0
flask-sqlalchemy=2.5.1=pypi_0
frozenlist=1.3.0=pypi_0
future=0.18.2=pypi_0
googlemaps=4.6.0=pypi_0
greenlet=1.1.2=pypi_0
idna=3.3=pypi_0
importlib-metadata=4.12.0=pypi_0
importlib-resources=5.9.0=pypi_0
itsdangerous=2.1.2=pypi_0
jinja2=3.1.2=pypi_0
line-bot-sdk=2.3.0=pypi_0
mako=1.2.1=pypi_0
markupsafe=2.1.1=pypi_0
multidict=6.0.2=pypi_0
oauthlib=3.2.1=pypi_0
openssl=1.1.1q=h2bbff1b_0
pillow=9.2.0=pypi_0
pip=22.1.2=py38haa95532_0
psycopg2=2.9.3=pypi_0
pycparser=2.21=pypi_0
pyjwt=2.4.0=pypi_0
python=3.8.13=h6244533_0
python3-openid=3.2.0=pypi_0
qrcode=7.3.1=pypi_0
requests=2.28.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
setuptools=61.2.0=py38haa95532_0
sqlalchemy=1.4.39=pypi_0
sqlite=3.38.5=h2bbff1b_0
sqlparse=0.4.2=pypi_0
tzdata=2022.1=pypi_0
urllib3=1.26.11=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=2.2.1=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py38haa95532_2
yarl=1.7.2=pypi_0
zipp=3.8.1=pypi_0
